refactor: VaultSuggestions to connect directly to pipes
